requirements-test.txt
# Testing Dependencies for Cleanup Toolkit
# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1  # For parallel test execution

# Code coverage and reporting
coverage[toml]>=7.2.7
codecov>=2.1.13

# Mocking and fixtures
mock>=5.1.0
faker>=19.2.0
factory-boy>=3.3.0

# Testing utilities
responses>=0.23.1  # For mocking HTTP requests
freezegun>=1.2.2  # For mocking time
testfixtures>=7.1.0  # Test fixtures and helpers

# Code quality and linting for tests
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
pylint>=2.17.4
mypy>=1.4.1

# Security testing
bandit>=1.7.5
safety>=2.3.5

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0

# Shell script testing
bashtest>=0.0.8
shellcheck-py>=0.9.0.5

# Git testing utilities
GitPython>=3.1.31

# Documentation testing (removed invalid package)

# Test reporting
pytest-html>=3.2.0
pytest-json-report>=1.5.0
allure-pytest>=2.13.2

# Environment management
python-dotenv>=1.0.0
pyyaml>=6.0.1

# HTTP client for integration tests
requests>=2.31.0
httpx>=0.24.1

# File system utilities
pathlib2>=2.3.7.post1
watchdog>=3.0.0